requirements.txt
google-api-python-client==2.160.0
PyGitHub==2.5.0
python-dotenv==1.0.1
numpy==2.2.0
